Move view test into a new class, re #24531
